all bugs working

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.CategoryDao;
 import org.yearup.data.ProductDao;
 import org.yearup.data.mysql.MySqlCategoryDao;
@@ -42,8 +43,18 @@ public class CategoriesController {
     // add the appropriate annotation for a get action
     @GetMapping("/{id}")
     public Category getById(@PathVariable int id) {
-        // get the category by id
-        return categoryDao.getById(id);
+        try {
+            var category = categoryDao.getById(id);
+
+            if (category == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+            return category;
+        } catch (ResponseStatusException ex) {
+            throw ex; // preserve 404
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 
     // the url to return all products in category 1 would look like this
@@ -62,7 +73,11 @@ public class CategoriesController {
     @ResponseStatus(value = HttpStatus.CREATED)
     public Category addCategory(@RequestBody Category category) {
         // insert the category
-        return categoryDao.create(category);
+        try {
+            return categoryDao.create(category);
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Sorry");
+        }
     }
 
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
@@ -83,7 +98,15 @@ public class CategoriesController {
     @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteCategory(@PathVariable int id) {
-        categoryDao.delete(id);
-        // delete the category by id
+        {
+            try {
+                var category = categoryDao.getById(id);
+                if (category == null)
+                    throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+                categoryDao.delete(id);
+            } catch (Exception ex) {
+                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+            }
+        }
     }
 }
